修改rdma_server的返回值为ssize_t
* rdma_server 的返回值如果大于零，就表示本次 RDMA 通信中，从端被修改的内存区域大小。 int 类型最大的修改量为2GB，不适用于文件或数据量表示，因此将其修改为ssize_t类型。这样不仅能通过返回值是否小于零判断函数调用是否成功，还足够表示绝大多数文件大小。

+++ kvstore.h
@@ -129,7 +129,7 @@ int kvs_slave_sync();
 int kvs_master_full_sync(client_info * cli);
 int kvs_master_incr_sync(client_info * cli, char ** tokens, int count);
 int rdma_client(const char * server_ip, const char * port, char * ptr, size_t size);
-int rdma_server(const char * port, char * rdma_buf, size_t size, int sockfd);
+ssize_t rdma_server(const char * port, char * rdma_buf, size_t size, int sockfd);
 // kvs_sync.c
 
 int kvs_config_init(kvs_conf_t *  conf);
